EDCA/requirements.txt
albumentations==1.4.2
colorama==0.4.6
matplotlib==3.6.2
numpy==1.26.4
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
pandas==2.1.3
Pillow==9.3.0
scipy==1.12.0
